+++ src/main/java/net/silvertide/kindred/bond/bond_results/BreakResult.java
@@ -2,5 +2,6 @@ package net.silvertide.kindred.bond.bond_results;
 
 public enum BreakResult {
     BROKEN,
-    NO_SUCH_BOND
+    NO_SUCH_BOND,
+    RELEASE_FAILED
 }